refactor() : adapt imports and the call to certain existing functions that are moved to different files.
